# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agora API",
    "version": "0.1.0",
    "description": "Industrial-grade marketplace SaaS — API contracts (ADR-0011)."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/auth/refresh": {
      "post": {
        "operationId": "authRefresh",
        "summary": "Rotate refresh token (session family rotation)",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthRefreshRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "New token pair",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthLoginResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid/reused token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "authLogout",
        "summary": "Revoke a session",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLogoutRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Revoked"
          }
        }
      }
    },
    "/v1/auth/reset/request": {
      "post": {
        "operationId": "passwordResetRequest",
        "summary": "Request password reset (non-enumerating)",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted"
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/auth/reset/confirm": {
      "post": {
        "operationId": "passwordResetConfirm",
        "summary": "Confirm password reset with one-time token",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirmRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated; all sessions revoked"
          },
          "400": {
            "description": "Invalid token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "authLogin",
        "summary": "Login — access + refresh tokens (M1)",
        "tags": [
          "auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session established (or MFA challenge when required)",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AuthLoginResponse"
                    },
                    {
                      "$ref": "#/components/schemas/MfaChallengeResponse"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "423": {
            "description": "Account locked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Machine-readable error code, e.g. EMAIL_TAKEN, INVALID_CREDENTIALS",
                "example": "EMAIL_TAKEN"
              },
              "message": {
                "type": "string",
                "description": "Human-readable message",
                "example": "Email is already registered"
              },
              "requestId": {
                "type": "string",
                "description": "Correlation id (x-request-id)"
              },
              "details": {
                "type": "object",
                "additionalProperties": {},
                "description": "Optional field-level details"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      },
...
      "AuthRefreshRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32,
            "maxLength": 256
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "AuthLoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "RS256 JWT, 15 min TTL"
          },
          "refreshToken": {
            "type": "string",
            "description": "Opaque refresh token, 30 d TTL, rotated on use"
          },
          "expiresIn": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "description": "Access token TTL in seconds",
            "example": 900
          },
          "tokenType": {
            "type": "string",
            "const": "Bearer"
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "description": "UUID v7 identifier",
                "example": "01961e78-4d99-7c20-8000-000000000000"
              },
              "email": {
                "type": "string",
                "format": "email"
              },
              "mfaRequired": {
                "type": "boolean"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "id",
              "email",
              "mfaRequired",
              "roles"
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "tokenType",
          "user"
        ]
      },
      "AuthLogoutRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "minLength": 32,
            "maxLength": 256
          }
        }
      },
      "PasswordResetRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          }
        },
        "required": [
          "email"
        ]
      },
      "PasswordResetConfirmRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 32,
            "maxLength": 256
          },
          "newPassword": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
      "AuthLoginRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 320
          },
          "password": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "mfaCode": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6,
            "description": "TOTP code when MFA is enabled"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "MfaChallengeResponse": {
        "type": "object",
        "properties": {
          "mfaRequired": {
            "type": "boolean",
            "const": true
          },
          "challenge": {
            "type": "string",
            "description": "Short-lived MFA challenge token (2 min)"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "UUID v7 identifier",
            "example": "01961e78-4d99-7c20-8000-000000000000"
          }
        },
        "required": [
          "mfaRequired",
          "challenge",
          "userId"
        ]
      }
    }
  }
}